# Samples/Sample.Web.API/Web.API.Tests/Contracts/swagger-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Synergy sample API",
    "description": "<label>API Version</label>: <strong>1.0 </strong><br/> <label>Application Name</label>: <strong>Synergy Samples</strong><br/> <label>Application Version</label>: <strong>1.0.0.0</strong><br/> <label>Application Created on</label>: <strong>DEVELOPERS MACHINE</strong>",
    "contact": {
      "name": "Synergy software",
      "url": "https://github.com/synergy-software",
      "email": "synergy@todo.com"
    },
    "license": {
      "name": "Use under MIT License",
      "url": "https://github.com/synergy-software/net-api-best-practices/blob/master/LICENSE"
    },
    "version": "1.0"
  },
  "paths": {
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Returns all users stored in the system.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersQueryResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Creates a User.",
        "description": "Sample request:\r\n\r\n    POST /api/users\r\n    {\r\n       \"login\": \"marcin@synergy.com\",\r\n    }",
        "requestBody": {
          "description": "Details of user to create",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserCommand"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserCommandResult"
                }
              }
            }
          },
          "400": {
            "description": "If the request contains invalid data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "GetUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserQueryResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserReadModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "login": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "nullable": true
      },
      "GetUsersQueryResult": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserReadModel"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "nullable": true
      },
      "CreateUserCommand": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": true
          }
        },
        "nullable": true
      },
      "CreateUserCommandResult": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserReadModel"
          }
        },
        "nullable": true
      },
      "Object": {
        "type": "object",
        "nullable": true
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Object"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/Object"
        },
        "nullable": true
      },
      "GetUserQueryResult": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserReadModel"
          }
        },
        "nullable": true
      }
    }
  }
}